Move ProfileEdit and ProfileSummary components to the profile folder Styling changes to profile view
